renderer: fix leak in hashtable on replace
Some american fuzzy lop tests managed to replace resources, however the
old values got leaked (I am not sure if this should be allowed)

In all cases, introduce a destroy callback to the hashtable, used when a
value is removed, replaced or the table is cleared. This simplifies a
bit resource management and help avoid potential errors by using simply
refcounts.

Signed-off-by: Marc-André Lureau <marcandre.lureau@redhat.com>