renderer: remove winsys_adjust.zw (DepthRange scale, trans)
Previously, vrend_shader insterted a scale & bias to emulate
glDepthRange in all vertex shaders.  This approach will fail
to match the gl spec in some situations.   glDepthRange() is also
called, causing the transformation to be applied twice.  The
winsys_adjust uniform is now a scalar to implement the y-flip only.

This bug was discovered and tested using the
dEQP-GLES2.functional.depth_range test suite.  This test now passes
all but one case, which appears to be a separate issue.

Signed-off-by: Joe M. Kniss <djmk@google.com>
[airlied: fixes some piglit tests as well - no regressions]
Signed-off-by: Dave Airlie <airlied@redhat.com>