vrend: Warn on different type and format when reading back FBOs v2
Warn if the expected native format/type differs from the allowed
type and format that are accepted by the OpenGL ES spec.

This is hopefully only a problem for Mesa which we have now fixed.
Other option is to add two extra copies and format conversions.

v2: More verbose and better comment.

Reviewed-by: Tomeu Vizoso <tomeu.vizoso@collabora.com>
Signed-off-by: Jakob Bornecrantz <jakob@collabora.com>