vrend: enable multisample on GLES host with GLES >= 3.0
Multisample fragment operations are part of the OpenGL ES 3.0 spec and
enabling them doesn't required the extensions to be listed explicitly.

Fixes:

 dEQP-GLES3.functional.multisample.fbo_max_samples.proportionality_alpha_to_coverage
 dEQP-GLES3.functional.multisample.fbo_max_samples.proportionality_sample_coverage
 dEQP-GLES3.functional.multisample.fbo_max_samples.proportionality_sample_coverage_inverted
 dEQP-GLES3.functional.multisample.fbo_max_samples.sample_coverage_invert

on an GLES host with GLES >= 3.0.

v2: - Do not set GL_SAMPLE_ALPHA_TO_ONE state when on gles host, because it is
      not supported there. This silences a number of according warnings on the
      host.

v3: - Properly place if statement to not send GL_SAMPLE_ALPHA_TO_ONE.
    - Do not set GL_MULTISAMPLE state when on gles host (same as above).

v4: - fix nitpicks in commit message

v5: - seperate setting of GL_MULTISAMPLE (not available on GLES) and
      GL_SAMPLES_TEST

Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>
Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Signed-off-by: Jakob Bornecrantz <jakob@collabora.com>