vrend: Call glDepthRange if we move from default value
Virglrender holds the values in a structure vrend_sub_context that is zeroed
out when it is initialized.
So, when the first state change set the near_val and far_val to zero,
the check assumes it is the old state and doesn't call glDepthRange.

Fixes: dEQP-GLES2.functional.depth_range.write.zero_to_zero

v2: initialize far_val in vrend_renderer_create_sub_ctx (Gurchetan)

Suggested-by: Gert Wollny <gert.wollny@collabora.com>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>
Signed-off-by: Elie Tournier <elie.tournier@collabora.com>
Signed-off-by: Jakob Bornecrantz <jakob@collabora.com>