virgl_hw.h: use same license statement as virgl_protocol.h
There's a few minor mistakes in the old license statement, for
instance:
- missing the year 2014 from the copyright
- "without limitation the" instead of "without limitation on the"

Apart from that it's just minor details like wrapping, and adding
"(including the next paragraph)". None of this seems to change
anything from a legal standpoint.

This matches the copyright-header we have for the same file in
Mesa, and makes it a bit less noisy to update the file there.

Signed-off-by: Erik Faye-Lund <erik.faye-lund@collabora.com>
Acked-by: Dave Airlie <airlied@redhat.com>