vrend: use the row-stride when directly reading back to an IOV
When guest mesa does a glReadPixel call, it expects the data to be placed
with correct row strides into the readback buffer. While for calls that
invert or require more then one IOV this is done in write_transfer_data,
for a read that only uses a small area and offset making it fit into the
first IOV this was not done. Consequently, in this case configure the
PACK_ROW_LENGTH accordingly.

v2: fix whitespace issues

This fixes a number of piglits:
   arb_copy_image-targets (for texture_2d source or dest, non-compressed)
   arb_get_texture_multisample/sample_position
   arb_texture_rectangle/fbo-blit rect
   arb_get_texture__sub_image-cubemap

Reviewed-by: Elie Tournier <elie.tournier@collabora.com>
Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Signed-off-by: Jakob Bornecrantz <jakob@collabora.com>