protect gl{Begin, End}ConditionalRendering calls
Otherwise, it'd be possible to generate evil commands from a rouge
guest-driver that can crash the VM.

This is a bit trickier than the previous one, because we were already
mixing calls to the OpenGL 3.0 version and the GL_NV_conditional_render
version, which indicates that this was previously only safe if *both*
were supported, that is OpenGL 3.0 *with* GL_NV_conditional_render.

Now, this code should match the caps we generate, which shouldn't give
any percieved feature-regressions or gl-versions supported.

v2: do not set have_gl_conditional_render if on GLES 3.0 or above.

Signed-off-by: Erik Faye-Lund <erik.faye-lund@collabora.com>
Tested-by: Po-Hsien Wang <pwang@chromium.org>
Signed-off-by: Dave Airlie <airlied@redhat.com>