features: convert current feature list to an array (v2)
Instead of having the features in separate named bool, convert it
to an array of bools indexed by an enum.

This is a straight mechanical patch, no code flows should change
and it doesn't change any functionality, it just refactors the
code to use the enum list and a bool array.

Future work should be a lot easier on this, so I'd like to
land it first.

v2: use lowercase enum, move debug_cb in as well.
rename tf2, sort list.

Reviewed-by: Elie Tournier <elie.tournier@collabora.com>
Reviewed-By: Gert Wollny <gert.wollny@collabora.com>