discourage using legacy-definitions
This copy of the bind-flags should not be used, as they are a duplicate
set of what can already be found in virgl_hw.h.

But to avoid breaking theoretical, unknown users, we shouldn't remove
them either. So let's just add a comment to tell people not to use
them.

Signed-off-by: Erik Faye-Lund <erik.faye-lund@collabora.com>
Reviewed-by: Dave Airlie <airlied@redhat.com>