shader: emit precision for images on GLSL
When using OpenGL ES 3.1, we either need to emit a precision qualifier,
or set a default precision.

Since it's not unthinkable that we'd want to forward lowp/mediump
in the future, let's use the precision qualifier so it's easier
to get this right later on.

This gets rid of a shader compile error.

Signed-off-by: Erik Faye-Lund <erik.faye-lund@collabora.com>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>
Signed-off-by: Dave Airlie <airlied@redhat.com>