shader: Generate arrays only after the declarations are parsed.
When we enable indiviudal IO arrays that are passed from the host we
will no longer generate the one big array, but we know only after all
declarations have been parsed, whether the guest emitted these arrays or
whether we have indirect access to elements but no arrays so that we
need to build them. So do this array generation after parsing the
declarations.

v2: - If available use the output ranges size of the previous shader when
      defining the input range for the generated arrays, mesa might have
      optimized away some individual values.
    - Add a GLSL version requirement when an interface block will be created.

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org> (v1)