vrend, blit: Use GL 4.4 logic for blitting
On GL 4.4 (and GLES with EXT_sRGB_write_control) blits will decode and
encode sRGB when GL_FRAMEBUFFER_SRGB is supported, and they don't decode
when it isn't supported. Enabling or disabling this is signalled from
the guest by using the according surfaces.

In summary, there is no need to use the GL fallback for a blit decoding
sRGB, because this can also be handled in a normal blit.

Fixes piglits on a GL host:
  blit texture srgb_to_linear downsample enabled render
  blit renderbuffer srgb_to_linear msaa enabled render
  blit renderbuffer srgb_to_linear msaa enabled clear
  blit renderbuffer srgb_to_linear downsample enabled clear

v2: Fix typo in commit message (Erik)

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>