vrend,blit: Fix the use of glCopyTexSubImage
On one hand render_condition_enable might also be set when the
cond_render_gl_mode is empty, so no conditional rendering is
actually done and the copy path can be used. On the other hand,
on GLES rectangular textures are emulated, so the texture target
needs to be corrected, and for some reason different base formats
were required which doesn't make sense.

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>