vrend: Don't expect TCS to be present in first shader pass
The shaders are issued backwards starting from the fragment
shader and, therefore, in the first pass when we fill out
the shader keys for the TES no TCS will be present.

So don't bail out in this case and assume there is a VS.
Later, when the VS shader is compiled it will check whether
a TES is present without a TCS and report the error.

Fixes: 956a6ceb8d8ffe3883277c75eaf5b4f48222ff68
shader: Pass information about the layout of generics and patches to the next stage

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Reviewed-by: <Gurchetan Singh gurchetansingh@chromium.org>