vrend,shader: Inject a TC shader if it is missing on GLES
On D-GL a program may contain a TE shader but no TC shader. On  GLES either
both or none of TES and TCS need to be available. So if the guest sends a
shader program without a TCS, inject a passthrough shader using the
patch parameters given in the GL code.

Closes #84

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>